Move classes to own file
